tests: keep Win32 host resize test on Windows
