+++ win32_gui_test.go
@@ -72,20 +72,6 @@ func TestWin32GuiRenderer_CursorDirty(t *testing.T) {
 	}
 }
 
-func TestWin32GuiHost_ResizeGridWithoutWindow(t *testing.T) {
-	host := &Win32GuiHost{cols: 80, rows: 25, cellW: 8, cellH: 16}
-	host.ResizeGrid(120, 40)
-	if host.cols != 120 || host.rows != 40 {
-		t.Fatalf("logical grid = %dx%d, want 120x40", host.cols, host.rows)
-	}
-
-	host.ResizeGrid(0, 40)
-	host.ResizeGrid(120, 0)
-	if host.cols != 120 || host.rows != 40 {
-		t.Fatalf("invalid resize changed logical grid to %dx%d", host.cols, host.rows)
-	}
-}
-
 func TestWin32GuiRenderer_ImplementsInterfaces(t *testing.T) {
 	var _ SurfaceRenderer = (*Win32GuiRenderer)(nil)
 	var _ GraphicsRenderer = (*Win32GuiRenderer)(nil)
